refactor: move openclaw image turns to adapter
